==> 4_modeling.ipynb <==
import numpy as np
import pandas as pd
from sklearn.tree import DecisionTreeClassifier
from sklearn.datasets import make_blobs
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import cross_val_score, StratifiedKFold
from sklearn.model_selection import cross_validate
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier, BaggingClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.linear_model import Lasso, LassoCV
from sklearn.preprocessing import StandardScaler

reddit = pd.read_csv('../data/reddit_data_cleaned.csv')
df_tvec_10 = pd.read_csv('../data/title_tfidf_10.csv')
df_tvec_25 = pd.read_csv('../data/title_tfidf_25.csv')
df_tvec_50 = pd.read_csv('../data/title_tfidf_50.csv')
df_tvec_100 = pd.read_csv('../data/title_tfidf_100.csv')

#Setting up a few dummy variables

reddit = pd.get_dummies(reddit, columns=['posted_hour'])
reddit = pd.get_dummies(reddit, columns=['posted_weekday'])
# reddit = pd.get_dummies(reddit, columns=['subreddit'])

#merge the dataframes
reddit_tfidf_10 = reddit.merge(df_tvec_10, on='post_id', how='left')
#drop redundant columns
reddit_tfidf_10.drop(columns = ['stemmed_title'], inplace = True)

#merge the dataframes
reddit_tfidf_25 = reddit.merge(df_tvec_25, on='post_id', how='left')
#drop redundant columns
reddit_tfidf_25.drop(columns = ['stemmed_title'], inplace = True)

#merge the dataframes
reddit_tfidf_50 = reddit.merge(df_tvec_50, on='post_id', how='left')
#drop redundant columns
reddit_tfidf_50.drop(columns = ['stemmed_title'], inplace = True)

#merge the dataframes
reddit_tfidf_100 = reddit.merge(df_tvec_100, on='post_id', how='left')
#drop redundant columns
reddit_tfidf_100.drop(columns = ['stemmed_title'], inplace = True)

# Model Baseline Accuracy

len(reddit[reddit['target'] == 1])/len(reddit)

The baseline accuracy for my model is 50.14% (subject to change as I collect more data), if we predict that all posts are successful since there are slightly more successful posts that failed posts. For context, succesful is defined as getting above the median number of comments.

# RF Model Without TFIDF Terms

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'stemmed_title','subreddit']

#creating our X and y
X = reddit.drop(columns = drop)
y = reddit['target']

#random forest model with 100 estimators
my_forest = RandomForestClassifier(class_weight='balanced',
                                  n_estimators=100, 
                                  max_depth=10,
                                  min_samples_split=50,                        
                                  n_jobs=-1,
                                  random_state = 48)

score = cross_val_score(my_forest, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

After playing around with the parameters, I settled on a model that I believe will be less sususceptible to overfitting, with a relatively large number of min_samples_split and limiting max depth to 10. 

This model has a score of 0.71, with a 95% CI of between 0.73 and 0.68

However, if we are asking the question: 'how do you make a successful reddit post' then it feels disingenuous to include upvotes as an independent variable, because as a content creator you have no control over the number of upvotes that your post is receiving. 

#RF model excluding 'upvotes' from X

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'stemmed_title','subreddit', 'upvotes']

#creating our X and y
X = reddit.drop(columns = drop)
y = reddit['target']

#random forest model with 100 estimators
my_forest = RandomForestClassifier(class_weight='balanced',
                                  n_estimators=100, 
                                  max_depth=10,
                                  min_samples_split=50,                        
                                  n_jobs=-1,
                                  random_state = 48)

score = cross_val_score(my_forest, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

Unsuprisingly, our models performance decreases significantly. We are now looking at a score of 0.57 with a 95% CI of between 0.53 and 0.62

# RF Model Including TFIDF Terms

#RF model with top 10 TFIDF terms 

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'subreddit', 'upvotes']

#creating our X and y
X = reddit_tfidf_10.drop(columns = drop)
y = reddit_tfidf_10['target']

my_forest = RandomForestClassifier(class_weight='balanced',
                                  n_estimators=100, 
                                  max_depth=10,
                                  min_samples_split=50,                        
                                  n_jobs=-1,
                                  random_state=48)

score = cross_val_score(my_forest, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

Here we see that including the TFIDF terms does not really change the performance of our model. The CI narrows ever so slightly. Next we will check what happens when including more TFIDF terms

#RF model with top 100 TFIDF terms

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'subreddit', 'upvotes']

#creating our X and y
X = reddit_tfidf_100.drop(columns = drop)
y = reddit_tfidf_100['target']

my_forest = RandomForestClassifier(class_weight='balanced',
                                  n_estimators=100, 
                                  max_depth=10,
                                  min_samples_split=50,                        
                                  n_jobs=-1,
                                  random_state=48)

score = cross_val_score(my_forest, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

Again, we see that including more TFIDF terms does not improve the performance of our model. I believe this to be due to the fact that there are so many different subreddits, with so many different focuses, and titles are very short. On average, the titles in our data set are only 10 words long. For this reason, I will be choosing to exclude titles from my random forest model. 

# Logistic Regression Models

#logistic Regression without TFIDF terms

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'stemmed_title', 'subreddit', 'upvotes']

#creating our X and y
X = reddit.drop(columns = drop)
y = reddit['target']

logreg = LogisticRegression(n_jobs=-1)
score = cross_val_score(logreg, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

#logistic regression with TFIDF terms

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'subreddit', 'upvotes']

#creating our X and y
X = reddit_tfidf_10.drop(columns = drop)
y = reddit_tfidf_10['target']

logreg = LogisticRegression(n_jobs=-1)
score = cross_val_score(logreg, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

#logistic regression with TFIDF terms

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'subreddit', 'upvotes']

#creating our X and y
X = reddit_tfidf_100.drop(columns = drop)
y = reddit_tfidf_100['target']

logreg = LogisticRegression(n_jobs=-1)
score = cross_val_score(logreg, X, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

Our logistic model has a score of 0.57 with a 95% CI of between 0.52 and 0.62.

Similar to our random forest model, we see that our model's performance does not change significantly with the top 10 TFIDF terms, and worsens slightly with the top 100. I see no reason to include them given the current state of our model. This indicates that, given our data set, there are not any specific key words to include that will improve or hurt your chances of having a successful post. I believe this would change if we looked at data across a wider time frame, as posts related to current events are more likely to succeed, and all of our data was scraped over the course of a few weeks around the winter holidays. 

# Lasso Regression (Logistic)

# Lasso Regression with standard scaled variables

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'stemmed_title', 'subreddit', 'upvotes']

#creating our X and y
X = reddit.drop(columns = drop)
y = reddit['target']

#scaling the data
sc = StandardScaler()
Z = sc.fit_transform(X)

#creating a list of possible alphas
l_alphas = np.logspace(-10, 0, 50)

#testing for the optimal alpha
lasso_cv = LassoCV(alphas = l_alphas, cv = 5, max_iter = 50000)

lasso_cv.fit(Z, y)

var_coefs = list(zip(X.columns, list(lasso_cv.coef_)))

#creating a list of variables to drop
to_drop = []
for item_pair in var_coefs:
    if item_pair[1] == 0:
        to_drop.append(item_pair[0])


to_drop

Our lasso regression model zeros out several variables that do not have a significant impact on a post's chances at suceeding. Namely we see that whether or not a post contains a video does not have a significant impact, a list of hours that do not have a significant impact, and that monday and wednesday do not have a significant impact. 

# Final Logistic Model

#creating a regression with the reduced dataframe

#columns to drop for modeling
drop = ['title', 'hours_ago', 'scraped_time', 'target', 'posted_time', 
        'post_id', 'comments', 'stemmed_title', 'subreddit', 'upvotes']

#creating our X and y

X = X.drop(columns = to_drop)
sc = StandardScaler()
Z = sc.fit_transform(X)

logreg = LogisticRegression()
score = cross_val_score(logreg, Z, y)

score = cross_val_score(logreg, Z, y)
print(np.mean(score)+2*np.std(score), np.mean(score)-2*np.std(score))
print(np.mean(score))

Our final model has a score of 0.57 with a 95% CI of between 0.53 and 0.62

cross_validate(logreg, Z, y)

#selecting a model to examine the coefficients
#we select the third model as it has the score closest to the mean
final_results = cross_validate(logreg, Z, y, return_estimator=True)
final_logreg = final_results['estimator'][2]

coefficient_list = list(zip(X.columns, final_logreg.coef_[0]))
coefficient_df = pd.DataFrame(zip(X.columns, final_logreg.coef_[0]))
coefficient_df = coefficient_df.rename(columns = {0:'Variable',1:'Coefficient'})
coefficient_df.sort_values('Coefficient', ascending = False).head(20)

Examining our coefficients, we can determine that to give our post the best chance of succeeding, it should be a text only post, posted on a thursday, around 10-11 am. The worst things we could do for our chances of having a successful post would be posting at an image on a friday between 11pm and midnight.

It should be noted that our final model does not have a lot of predictive power, but ultimately what we are trying to predict is human behaviour. We are trying to predict what people like, and how they will interact with a post. When the scope of our analysis is the front page, it is very difficult to determine what will succeed. I don't even have a strong understanding of the algorithm used to determine what ends up on the front page. This project would likely be much more successful if we took a deep dive into a single subreddit, and tried to determine what attributes helped a post become successful within the boundaries of that subreddit. 

If we were to narrow the scope of our analysis in this way, I believe that the TFIDF terms would become much more powerful predictors. For example, in the anti-work subreddit, I would imagine that posts with much more negative titles would tend to do better, as the people that are members of that subreddit love to see people talking trash about their bosses and explaining their unpleasant working conditions.
